🐛 fix: pR 리뷰 반영 코드 수정
